# docs/swagger.yaml
basePath: /
definitions:
  email-service_internal_model.Account:
    description: Holds credentials and configuration for an external email service.
    properties:
      access_key_id:
        type: string
      created_at:
        type: string
      daily_send_limit:
        type: integer
      domain:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handler.CreateAccountRequest:
    description: Request body for creating a new Aliyun email account.
    properties:
      access_key_id:
        example: LTAI5txxxxxxxxxxxxxx
        type: string
      access_key_secret:
        example: P1aXxxxxxxxxxxxxxxxxxxxxxx
        type: string
      daily_send_limit:
        example: 5000
        type: integer
      domain:
        example: mail.example.com
        type: string
      name:
        example: Aliyun Account 1
        type: string
    required:
    - access_key_id
    - access_key_secret
    - domain
    - name
    type: object
...
  internal_handler.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
  internal_handler.SendEmailRequest:
    properties:
      account_sender_id:
        example: 1
        type: integer
      html_body:
        example: <p>Hello, this is a test email.</p>
        type: string
      subject:
        example: A Quick Message
        type: string
      to_address:
        example: test.recipient@example.com
        type: string
    required:
    - account_sender_id
    - html_body
    - subject
    - to_address
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: email-service/internal/handler
    url: http://www.swagger.io/support
  description: This is a powerful email service API for batch sending and management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Email Service API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      description: Retrieves a list of all configured email service provider accounts.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/email-service_internal_model.Account'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get All Accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Adds a new third-party email service provider account.
      parameters:
      - description: Account Details
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Account'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Account
      tags:
      - Accounts
  /api/v1/accounts/{id}:
    delete:
      description: Deletes an account by its ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Account deleted successfully'
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid account ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Account
      tags:
      - Accounts
    get:
      description: Retrieves the details of a specific account by its ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Account'
              type: object
        "400":
          description: Invalid account ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Account by ID
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: Updates the details of an existing account. The request body should
        contain the full account object.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account object with updated details
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/email-service_internal_model.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Account'
              type: object
        "400":
          description: Invalid account ID or request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Account
      tags:
      - Accounts
  /api/v1/emails/send:
    post:
      consumes:
      - application/json
      description: Queues a single, non-template email for immediate sending. This
        is suitable for transactional emails.
      parameters:
      - description: Email Details
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SendEmailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: 'message: Email queued for sending, record_id: 12345'
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Send Single Email
      tags:
      - Emails
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"